New RenderedTextList, which JournalGump uses.
And now I can reuse RenderedTextList for VendorBuyGump!
