+++ dev/Ultima/UI/Controls/HtmlGumpling.cs
@@ -9,14 +9,12 @@
  *
  ***************************************************************************/
 #region using
-
 using System;
 using Microsoft.Xna.Framework;
 using UltimaXNA.Core.Graphics;
 using UltimaXNA.Core.Input.Windows;
 using UltimaXNA.Core.UI;
 using UltimaXNA.Core.UI.HTML;
-using UltimaXNA.Ultima.UI.Interfaces;
 #endregion
 
 namespace UltimaXNA.Ultima.UI.Controls
