+++ dev/Ultima/UI/Controls/ScrollFlag.cs
@@ -14,7 +14,6 @@ using UltimaXNA.Core.Graphics;
 using UltimaXNA.Core.Input.Windows;
 using UltimaXNA.Core.UI;
 using UltimaXNA.Ultima.IO;
-using UltimaXNA.Ultima.UI.Interfaces;
 #endregion
 
 namespace UltimaXNA.Ultima.UI.Controls
@@ -136,7 +135,7 @@ namespace UltimaXNA.Ultima.UI.Controls
             }
             else
             {
-                spriteBatch.Draw2D(m_GumpSlider, new Vector3(position.X, position.Y + m_SliderPosition, 0), Vector3.Zero);
+                spriteBatch.Draw2D(m_GumpSlider, new Vector3(position.X - 5, position.Y + m_SliderPosition, 0), Vector3.Zero);
             }
 
             base.Draw(spriteBatch, position);
@@ -160,6 +159,7 @@ namespace UltimaXNA.Ultima.UI.Controls
 
         protected override bool IsPointWithinControl(int x, int y)
         {
+            x -= 5;
             Rectangle slider = new Rectangle(0, (int)m_SliderPosition, m_GumpSlider.Width, m_GumpSlider.Height);
             return slider.Contains(x, y);
         }
