+++ dev/Ultima/UI/WorldGumps/JournalGump.cs
@@ -22,8 +22,8 @@ namespace UltimaXNA.Ultima.UI.WorldGumps
     class JournalGump : Gump
     {
         private ExpandableScroll m_Background;
-        private List<RenderedText> m_JournalEntries;
-        private ScrollBar m_ScrollBar;
+        private RenderedTextList m_JournalEntries;
+        private IScrollBar m_ScrollBar;
 
         public JournalGump()
             : base(0, 0)
@@ -33,111 +33,32 @@ namespace UltimaXNA.Ultima.UI.WorldGumps
             AddControl(m_Background = new ExpandableScroll(this, 0, 0, 300));
             m_Background.TitleGumpID = 0x82A;
 
-            AddControl(m_ScrollBar = new ScrollBar(this));
-            m_ScrollBar.IsVisible = false;
+            m_ScrollBar = (IScrollBar)AddControl(new ScrollFlag(this));
+            AddControl(m_JournalEntries = new RenderedTextList(this, 30, 36, 242, 200, m_ScrollBar));
         }
 
         protected override void OnInitialize()
         {
             SetSavePositionName("journal");
 
-            m_JournalEntries = new List<RenderedText>();
             InitializeJournalEntries();
             PlayerState.Journaling.OnJournalEntryAdded += AddJournalEntry;
         }
 
         public override void Update(double totalMS, double frameMS)
         {
+            m_JournalEntries.Height = Height - 98;
             base.Update(totalMS, frameMS);
-
-            m_ScrollBar.Position = new Point(Width - 45, 35);
-            m_ScrollBar.Height = Height - 100;
-            CalculateScrollBarMaxValue();
-            m_ScrollBar.IsVisible = m_ScrollBar.MaxValue > m_ScrollBar.MinValue;
         }
 
         public override void Draw(SpriteBatchUI spriteBatch, Point position)
         {
             base.Draw(spriteBatch, position);
-
-            Point p = new Point(position.X + 36, position.Y + 35);
-            int height = 0;
-            int maxheight = m_ScrollBar.Value + m_ScrollBar.Height;
-
-            for (int i = 0; i < m_JournalEntries.Count; i++)
-            {
-                if (height + m_JournalEntries[i].Height <= m_ScrollBar.Value)
-                {
-                    // this entry is above the renderable area.
-                    height += m_JournalEntries[i].Height;
-                }
-                else if (height + m_JournalEntries[i].Height <= maxheight)
-                {
-                    int y = height - m_ScrollBar.Value;
-                    if (y < 0)
-                    {
-                        // this entry starts above the renderable area, but exists partially within it.
-                        m_JournalEntries[i].Draw(spriteBatch, new Rectangle(p.X, position.Y + 35, m_JournalEntries[i].Width, m_JournalEntries[i].Height + y), 0, -y);
-                        p.Y += m_JournalEntries[i].Height + y;
-                    }
-                    else
-                    {
-                        // this entry is completely within the renderable area.
-                        m_JournalEntries[i].Draw(spriteBatch, p);
-                        p.Y += m_JournalEntries[i].Height;
-                    }
-                    height += m_JournalEntries[i].Height;
-                }
-                else
-                {
-                    int y = maxheight - height;
-                    m_JournalEntries[i].Draw(spriteBatch, new Rectangle(p.X, position.Y + 35 + m_ScrollBar.Height - y, m_JournalEntries[i].Width, y), 0, 0);
-                    // can't fit any more entries - so we break!
-                    break;
-                }
-            }
-        }
-
-
-        private void CalculateScrollBarMaxValue()
-        {
-            bool maxValue = m_ScrollBar.Value == m_ScrollBar.MaxValue;
-
-            int height = 0;
-            for (int i = 0; i < m_JournalEntries.Count; i++)
-            {
-                height += m_JournalEntries[i].Height;
-            }
-
-            height -= m_ScrollBar.Height;
-
-            if (height > 0)
-            {
-                m_ScrollBar.MaxValue = height;
-                if (maxValue)
-                    m_ScrollBar.Value = m_ScrollBar.MaxValue;
-            }
-            else
-            {
-                m_ScrollBar.MaxValue = 0;
-                m_ScrollBar.Value = 0;
-            }
         }
 
         private void AddJournalEntry(string text)
         {
-            bool maxScroll = (m_ScrollBar.Value == m_ScrollBar.MaxValue);
-
-            while (m_JournalEntries.Count > 99)
-            {
-                m_JournalEntries.RemoveAt(0);
-            }
-            m_JournalEntries.Add(new RenderedText(string.Format("<left color='#50422D'><span width='14'/>{0}</left><br/>", text), 200));
-            m_ScrollBar.MaxValue += m_JournalEntries[m_JournalEntries.Count - 1].Height;
-            if (maxScroll)
-            {
-                m_ScrollBar.Value = m_ScrollBar.MaxValue;
-            }
+            m_JournalEntries.AddEntry(string.Format("<left color='#50422D'><span width='14'/>{0}</left><br/>", text));
         }
 
         private void InitializeJournalEntries()
