+++ dev/UltimaXNA.csproj
@@ -204,6 +204,7 @@
     <Compile Include="Core\UI\Interfaces\IFont.cs" />
     <Compile Include="Core\UI\Interfaces\IUIResourceProvider.cs" />
     <Compile Include="Core\PluginManager.cs" />
+    <Compile Include="Core\UI\RenderedTextList.cs" />
     <Compile Include="ServiceRegistry.cs" />
     <Compile Include="Ultima\Audio\AudioService.cs" />
     <Compile Include="Ultima\Audio\UOMusic.cs" />
@@ -307,7 +308,7 @@
     <Compile Include="Ultima\World\Entities\Mobiles\MobileFlags.cs" />
     <Compile Include="Core\UI\Fonts\AFont.cs" />
     <Compile Include="Ultima\Network\Server\FeatureFlags.cs" />
-    <Compile Include="Ultima\UI\Interfaces\IScrollBar.cs" />
+    <Compile Include="Core\UI\Interfaces\IScrollBar.cs" />
     <Compile Include="Ultima\UI\Controls\ScrollFlag.cs" />
     <Compile Include="Ultima\UI\Controls\CheckBox.cs" />
     <Compile Include="Ultima\UI\Controls\RadioButton.cs" />
